WIP: Basic inline assembly for proposed RVV zip
To show how the vector zips currently proposed for RVV can be used in
some code bases, add some implementations of the single-precision
copying (with reordering) of matrices ready for GEMM.

The inline assembly is rough and at least the inline part of this can be
improved for better performance. The intention for this commit is to
show how each of the different reorderings can make use of the proposed
instructions.